requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gensim==4.3.2
gensim-evaluations==0.1.2
gitdb==4.0.11
GitPython==3.1.40
h5py==3.10.0
harte-library==0.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jams==0.3.4
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lark==1.1.8
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mido==1.3.0
mir-eval==0.7
mirdata==0.3.7
mistune==3.0.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
music21==8.3.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
pretty-midi==0.2.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyrdf2vec==0.2.3
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-Levenshtein==0.12.2
python-louvain==0.15
pytorch-lightning==1.8.4
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
rdflib==7.0.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SPARQLWrapper==2.0.0
stack-data==0.6.3
suffix-tree==0.1.2
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.7.2
torch==1.12.0
torchmetrics==1.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wandb==0.15.12
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
Wikidata==0.7.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
